refactor: restructure SDK
