+++ src/api/externalUser.ts
@@ -1,17 +1,17 @@
-import { KeyManager, KeyManagerLevel } from '../sdk/services/keymanager';
-import { JWTLoginPayload, OnPressLoginOptions, UserApps } from '../sdk/userApps';
+import { KeyManager, KeyManagerLevel } from '../sdk/storage/keymanager';
+import { JWTLoginPayload, OnPressLoginOptions, UserApps } from '../sdk/controllers/userApps';
 import { createVCSigner, generateRandomKeyPair, randomString } from '../sdk/util/crypto';
 import { ES256KSigner } from '@tonomy/did-jwt';
 import { createJWK, toDid } from '../sdk/util/did-jwk';
-import { Message, MessageType } from '../sdk/util/message';
+import { Message, MessageType } from '../sdk/services/communication//message';
 import { getSettings } from '../sdk/settings';
-import { SdkErrors, throwError } from '../sdk/services/errors';
-import { createStorage, PersistentStorageClean, StorageFactory, STORAGE_NAMESPACE } from '../sdk/services/storage';
+import { SdkErrors, throwError } from '../sdk/util/errors';
+import { createStorage, PersistentStorageClean, StorageFactory, STORAGE_NAMESPACE } from '../sdk/storage/storage';
 import { Checksum256, Name } from '@greymass/eosio';
-import { TonomyUsername } from '../sdk/services/username';
-import { browserStorageFactory } from '../sdk/managers/browserStorage';
-import { getChainInfo } from '../sdk/services/eosio/eosio';
-import { JsKeyManager } from '../sdk/managers/jsKeyManager';
+import { TonomyUsername } from '../sdk/util/username';
+import { browserStorageFactory } from '../sdk/storage/browserStorage';
+import { getChainInfo } from '../sdk/services/blockchain/eosio/eosio';
+import { JsKeyManager } from '../sdk/storage/jsKeyManager';
 
 export type ExternalUserStorage = {
     accountName: Name;
