+++ src/sdk/controllers/app.ts
@@ -1,9 +1,9 @@
 /* eslint-disable camelcase */
 import { Name, PrivateKey, PublicKey } from '@greymass/eosio';
-import { IDContract } from './services/contracts/IDContract';
-import { createSigner } from './services/eosio/transaction';
-import { getSettings } from './settings';
-import { AccountType, TonomyUsername } from './services/username';
+import { IDContract } from '../services/blockchain/contracts/IDContract';
+import { createSigner } from '../services/blockchain/eosio/transaction';
+import { getSettings } from '../settings';
+import { AccountType, TonomyUsername } from '../util/username';
 
 const idContract = IDContract.Instance;
