+++ src/sdk/services/blockchain/eosio/authority.ts
