+++ src/sdk/services/blockchain/eosio/eosio.ts
@@ -1,8 +1,8 @@
 import { APIClient, FetchProvider } from '@greymass/eosio';
 import { GetInfoResponse } from '@greymass/eosio/src/api/v1/types';
 import fetch from 'cross-fetch';
-import { getSettings } from '../../settings';
-import { throwError, SdkErrors } from '../errors';
+import { getSettings } from '../../../settings';
+import { throwError, SdkErrors } from '../../../util/errors';
 
 let api: APIClient;
