+++ src/sdk/services/blockchain/eosio/transaction.ts
@@ -1,6 +1,6 @@
 import { Action, API, Transaction, SignedTransaction, Signature, Checksum256, Name, PrivateKey } from '@greymass/eosio';
-import { KeyManager, KeyManagerLevel } from '../keymanager';
-import { HttpError } from '../errors';
+import { KeyManager, KeyManagerLevel } from '../../../storage/keymanager';
+import { HttpError } from '../../../util/errors';
 import { getApi } from './eosio';
 
 type ActionData = {
