+++ src/sdk/storage/jsKeyManager.ts
@@ -6,11 +6,11 @@ import {
     GetKeyOptions,
     StoreKeyOptions,
     CheckKeyOptions,
-} from '../services/keymanager';
+} from './keymanager';
 import { Checksum256, PrivateKey, PublicKey, Signature } from '@greymass/eosio';
 import { createSigner } from '@tonomy/antelope-ssi-toolkit';
-import { SdkErrors, throwError } from '../services/errors';
-import { STORAGE_NAMESPACE } from '../services/storage';
+import { SdkErrors, throwError } from '../util/errors';
+import { STORAGE_NAMESPACE } from './storage';
 
 const KEY_STORAGE_NAMESPACE = STORAGE_NAMESPACE + 'key.';
