+++ test/services/keymanager.test.ts
@@ -2,8 +2,8 @@ import { Checksum256, PrivateKey } from '@greymass/eosio';
 import { JsKeyManager } from '../../src/sdk/managers/jsKeyManager';
 import argon2 from 'argon2';
 import { jsStorageFactory } from '../../src/cli/bootstrap/jsstorage';
-import { createUserObject } from '../../src/sdk/user';
-import { KeyManagerLevel } from '../../src/sdk/services/keymanager';
+import { createUserObject } from '../../src/sdk/controllers/user';
+import { KeyManagerLevel } from '../../src/sdk/storage/keymanager';
 import { randomBytes, generateRandomKeyPair } from '../../src/sdk/util/crypto';
 import { setSettings } from '../../src/sdk';
 import { generatePrivateKeyFromPassword } from '../../src/cli/bootstrap/keys';
